chore: rename main test package to metricbeat (its functional unit)
