+++ e2e/metricbeat/assertions.go
@@ -1,4 +1,4 @@
-package main
+package metricbeat
 
 import (
 	"fmt"
