+++ e2e/metricbeat/metricbeat_test.go
@@ -1,4 +1,4 @@
-package main
+package metricbeat
 
 import (
 	"context"
