+++ e2e/metricbeat/runner_test.go
@@ -1,4 +1,4 @@
-package main
+package metricbeat
 
 import (
 	"flag"
