+++ e2e/metricbeat/test_utils.go
@@ -1,4 +1,4 @@
-package main
+package metricbeat
 
 import (
 	"io"
